Dedup the tests
